fileleak/requirements.txt
lxml==4.4.1
python-docx==0.8.10
xlrd==1.2.0
